last case

+++ app/Listeners/MassLogsListener.php
@@ -60,47 +60,53 @@ class MassLogsListener
 
                 if($log->model == 'Enroll'){
 
-                    $user_old_announcements = Announcement::where('year_id',$one->year)
-                                                            ->where('type_id',$one->type)
-                                                            ->where('level_id',$one->level)
-                                                            ->where('class_id',$one->class)
-                                                            ->where('segment_id',$one->segment)
-                                                            ->where('course_id',$one->course)
-                                                            ->pluck('id');
-
-                    $user_old_announcements1 = AnnouncementsChain::where('year',$one->year)
-                                                            ->where('type',$one->type)
-                                                            ->where('level',$one->level)
-                                                            ->where('class',$one->class)
-                                                            ->where('segment',$one->segment)
-                                                            ->where('course',$one->course)
-                                                            ->pluck('announcement_id');
-
-                    $final_old_announcements = array_merge($user_old_announcements->toArray(),$user_old_announcements1->toArray());
-
-                    userAnnouncement::where('user_id',$one->user_id)->whereIn('announcement_id',$final_old_announcements)->delete();
-
-                    $notify = DB::table('notifications')->where('notifiable_id', $one->user_id)->get();
-                    $ids=collect();
-                    foreach ($notify as $not) {
-                        $not->data= json_decode($not->data, true);
-
-                        if($not->data['type'] == 'announcement' && in_array($not->data['id'],$final_old_announcements)){
-                            $ids->push($not->id);
-                        }
-
-                        if($not->data['type'] != 'announcement' && $not->data['course_id'] == $one->course){
-                            $ids->push($not->id);
-                        }
-                    }
-
-                    DB::table('notifications')->whereIn('id', $ids)->delete();
-
                     $user_enrolls = Enroll::where('user_id',$one->user_id)->where('id','!=',$one->id)->count();
                     if($user_enrolls == 0){
                         userAnnouncement::where('user_id',$one->user_id)->delete();
                         DB::table('notifications')->where('notifiable_id', $one->user_id)->delete();    
                     }
+
+                    if($user_enrolls != 0){
+
+                        $enrolls =  Enroll::where('user_id',$one->user_id)->where('id','!=',$one->id)->get();
+
+                        $user_old_announcements = Announcement::whereNotIn('year_id',$enrolls->pluck('year')->filter())
+                                                                ->orWhere('type_id',$enrolls->pluck('type')->filter())
+                                                                ->orWhere('level_id',$enrolls->pluck('level')->filter())
+                                                                ->orWhere('class_id',$enrolls->pluck('class')->filter())
+                                                                ->orWhere('segment_id',$enrolls->pluck('segment')->filter())
+                                                                ->orWhere('course_id',$enrolls->pluck('course')->filter())
+                                                                ->pluck('id');
+
+                        $user_old_announcements1 = AnnouncementsChain::whereNotIn('year',$enrolls->pluck('year')->filter())
+                                                                    ->orWhere('type',$enrolls->pluck('type')->filter())
+                                                                    ->orWhere('level',$enrolls->pluck('level')->filter())
+                                                                    ->orWhere('class',$enrolls->pluck('class')->filter())
+                                                                    ->orWhere('segment',$enrolls->pluck('segment')->filter())
+                                                                    ->orWhere('course',$enrolls->pluck('course')->filter())
+                                                                    ->pluck('id');
+
+                        $final_old_announcements = array_merge($user_old_announcements->toArray(),$user_old_announcements1->toArray());
+                        
+                        userAnnouncement::where('user_id',$one->user_id)->whereIn('announcement_id',$final_old_announcements)->delete();
+
+                        $notify = DB::table('notifications')->where('notifiable_id', $one->user_id)->get();
+                        $ids=collect();
+                        foreach ($notify as $not) {
+                            $not->data= json_decode($not->data, true);
+
+                            if($not->data['type'] == 'announcement' && in_array($not->data['id'],$final_old_announcements)){
+                                $ids->push($not->id);
+                            }
+
+                            if($not->data['type'] != 'announcement' && $not->data['course_id'] == $one->course){
+                                $ids->push($not->id);
+                            }
+                        }
+
+                        DB::table('notifications')->whereIn('id', $ids)->delete();
+                    }
+
                 }
             }
